requirements.txt
asgiref==3.10.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Django==5.2.7
et_xmlfile==2.0.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
whitenoise==6.11.0
